+++ plugins/modules/fortiflexvm_entitlements_vm_create.py
@@ -21,7 +21,7 @@ description:
     - This API is only used to create one or more VMs. To modify a VM, please refer to fortiflexvm_entitlements_update.
 version_added: "2.0.0"
 author:
-    - Xinwei Du (@DrMofu)
+    - Xinwei Du (@dux-fortinet)
 options:
     username:
         description: